schemaをexport defaultで解決できるように修正。テストコードにも最小限の変更
