+++ src/rmmz/plugin/schema/compile/kinds/core/x-rpg-param-schema.dataId.test.ts
@@ -1,15 +1,12 @@
 import { describe, test, expect } from "vitest";
 import Ajv from "ajv";
 import type { SourceIdentifier } from "src/namedItemSource";
-import type {
-  X_RmmzParam,
-  X_RmmzParamBase,
-} from "./core/paramBase/x-rpg-param";
-import type { X_ParamDataId } from "./x-rpg-param-schema";
-import { makeRpgParamMetaSchema } from "./x-rpg-param-schema";
+import type { X_RmmzParam, X_RmmzParamBase } from "./paramBase/x-rpg-param";
+import SCHEMA_RPG_PARAM_META_SCHEMA from "./x-rpg-param.schema";
+import type { X_ParamDataId } from "./x-rpg-paramTypes";
 
 const makeValidator = () => {
-  const schema = makeRpgParamMetaSchema();
+  const schema = SCHEMA_RPG_PARAM_META_SCHEMA;
   const ajv = new Ajv({
     discriminator: true,
     strict: true,
